### Summary
- Refactored multiple controllers to improve maintainability and performance.
- Added `User_FeatbackController` to handle user feedback feature.
- Updated models `DatPhong` and `UserFeedback` to support new business logic.
- Improved application configuration in `Program.cs` and `appsettings.json`.
- Removed unused files to optimize project structure.
- Updated build artifacts to reflect recent changes.

### Testing
- Verified all changes locally with successful tests.
- Ensured compatibility with the main branch.

### Next Steps
- Monitor feedback from the integration to ensure smooth operation.
- Plan for additional enhancements to the feedback system in future iterations.

+++ Models/UserFeedback.cs
@@ -9,7 +9,7 @@ public partial class UserFeedback
 
     public string MaPhong { get; set; } = null!;
 
-    public int SoSao { get; set; }
+    public float SoSao { get; set; }
 
     public string? NoiDung { get; set; }
